requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.8.6
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
